+++ test/nodes/test_connector.py
