+++ test/nodes/test_extractor.py
