+++ test/nodes/test_preprocessor.py
@@ -4,7 +4,7 @@ from haystack import Document
 from haystack.nodes.file_converter.pdf import PDFToTextConverter
 from haystack.nodes.preprocessor.preprocessor import PreProcessor
 
-from .conftest import SAMPLES_PATH
+from ..conftest import SAMPLES_PATH
 
 TEXT = """
 This is a sample sentence in paragraph_1. This is a sample sentence in paragraph_1. This is a sample sentence in
