+++ test/pipelines/test_pipeline_debug_and_validation.py
@@ -6,7 +6,7 @@ import pytest
 from haystack.pipelines import Pipeline, RootNode
 from haystack.nodes import FARMReader, BM25Retriever
 
-from .conftest import SAMPLES_PATH, MockRetriever as BaseMockRetriever, MockReader
+from ..conftest import SAMPLES_PATH, MockRetriever as BaseMockRetriever, MockReader
 
 
 class MockRetriever(BaseMockRetriever):
